Set dtype to make Windows happy

+++ tests/data_group_test.py
@@ -144,7 +144,7 @@ def test_numpy_arrays_are_not_considered_for_shape():
 
 def test_getitem_positional_indexing():
     dg = sc.DataGroup({'a': sc.arange('x', 4)})
-    assert sc.identical(dg['x', 2], sc.DataGroup({'a': sc.scalar(2)}))
+    assert sc.identical(dg['x', 2], sc.DataGroup({'a': sc.scalar(2, dtype='int64')}))
 
 
 def test_getitem_positional_indexing_leaves_scalar_items_untouched():
@@ -159,7 +159,7 @@ def test_getitem_positional_indexing_leaves_independent_items_untouched():
 
 def test_getitem_positional_indexing_without_dim_label_works_for_1d():
     dg = sc.DataGroup({'a': sc.arange('x', 4)})
-    assert sc.identical(dg[2], sc.DataGroup({'a': sc.scalar(2)}))
+    assert sc.identical(dg[2], sc.DataGroup({'a': sc.scalar(2, dtype='int64')}))
 
 
 def test_getitem_positional_indexing_without_dim_label_raises_unless_1d():
